Add styling, remove default Vue boilerplate css/templates
This commit removes the default Vue pages and re-organizes the components a little bit just to make the project more simple to read.
I also deleted all of the boilerplate templates/css that comes with it for the same reason. I also added a very small amount of CSS
just to make a live demo version of this nice to look at, but it's definitely not professional in any way.